Display EXECUTING_TASK messages in UI (#23411)
I think this is the best I can do inside the current layout/flow. Ideally we'd group the output by workspace (repository+path) and then show an overall status icon and the text without any prefixes, but I'm not sure what @courier-new is cooking up, so I merely decided to make this better for us.

As it is it allows me to quickly see what's happening and whether something went wrong.