stream results from searcher (#23754)
* rename Decoder to FrontendStreamDecoder

Renames Decoder to FrontendStreamDecoder to make way for a shared server sent
events decoder and a searcher-specific client.

* share a max payload size between client and server

Previously, if the writer wrote an event that was greater than the max
buffer size of the client, it would fail to parse and stop processing
further events. This can happen when we match a very large line, or have
many matches on a single line. Instead, if the writer shares the same
max event size, it can just ignore files matches that are too large, log
a message, and continue sending results.

* split out server sent event parser

This splits the server-sent event decoder into its own struct so it can
be re-used between the frontend stream client and the searcher stream
client.

* add StreamDecoder for searcher

* add callback and stream method to searcher client

Currently, this does nothing since the callback will always be nil, but
this will now read streamed events from searcher.

* move and export JSONArrayBuf

This makes JSONArrayBuf available to packages outside of cmd/frontend,
     and colocates it with the code that uses it.

* replace uses of limitedStreamCollector with an interface

This will make it easier to use a different streamer for the batch API
than for the streaming API.

* add limitedStream for a callback-based streamer

This adds a limitedStream type that implements matchSender. This lets us
use a callback to send results as they come rather than collecting them
into a slice as limitedStreamCollector does.

* add streaming support to searcher

This adds the ability for searcher to send results as it receives them
in the same server-sent-event format as we use when streaming from the
frontend to the UI.

This functionality is currently only activated when the client sends the
request with the URL query param Stream with the value "true".

* move match conversion code into a closure

This creates a closure `toMatches` which converts `protocol.FileMatch` into
`result.FileMatch` so it can be re-used between batch and streaming
codepaths.

* enable streaming end-to-end

This commit puts the last piece in place to enable streaming from end to
end. It passes a callback that is passed to `searcher.Search()` that,
when called, will forward streamed results down the stream in frontend.

* put streaming searcher behind a feature flag

* add decoder tests

* Update internal/search/streaming/http/writer.go

Co-authored-by: Rijnard van Tonder <rvantonder@gmail.com>

Co-authored-by: Rijnard van Tonder <rvantonder@gmail.com>