repo-updater: Add PermsSyncer to syncer & sync private repo perms (#24028)
In this commit we trigger a repo permissions sync for all private
repos (both added and modified) in the diff read from the
syncer.Synced channel. We do not make extra checks for existing
private repos that were modified as we do not anticipate a performance
based regression based on the number of repos being scheduled for
permissions syncing.