authz/github: sync user permissions from team/org perms caches (#23978)
Extend the GitHub provider's FetchUserPermsByToken to use cached groups (orgs or teams) permissions to reduce API calls required to do a full permissions sync. Also introduces options for performing permission syncs that can be provided at various levels of permissions syncing. For now, this is only added to user permissions, and is primarily used to provide an InvalidateCaches flag that signals the Provider to invalidate caches encountered during the sync, which is currently only used by webhook handlers to invalidate caches when a team/org-related event is received. The ScheduleUserPermissionsSync mutation and some webhook handlers also take advantage of this.

Co-authored-by: ᴜɴᴋɴᴡᴏɴ <joe@sourcegraph.com>