search: only show timed out alert if we don't show timedout progress (#24376)
Attempt yesterday didn't work. Easy to reproduce with very low timeouts
like 1ns. In that case repo resolution would fail, and we would return
the deadline as an error.

We now give a somewhat suboptimal solution since it is coupled to the
logic of how we decide if to show the timedout notification. However, I
wanna fix this and its my third attempt at trying to avoid that sort of
solution.