search: ignore dns lookup errors against zoekt (#24459)
During rollouts of Zoekt, we may still have endpoints of zoekt which are
not available in our endpoint map. In particular, this happens when
using Kubernetes and the (default) stateful set watcher.

This has the downside of incorrectly configured instances not relying on
service discovery always having an empty set. Right now this trade-off
is worth it because every rollout of Zoekt on sourcegraph.com results in
mostly unavailable search. To help offset this trade-off we instrument
metrics and tracing when we do this.

Co-authored-by: Stefan Hengl <stefan@sourcegraph.com>