authz/github: repo-centric perms sync from team/org perms caches (#24328)
Implements caching of groups permissions for repo-centric permissions sync. Follows up on the GitHub user-centric perms sync caching introduced in #23978. Repo sync now lists direct collaborators to a repo before querying for the organization and teams for users with access to this repo, and caching them as groups.

Also adds `allowGroupsPermissionsSync` to the GitHub OAuth provider config, which requests the additional `read:repo` scope required to enable `authorization.groupsCacheTTL` in GitHub code hosts.

Co-authored-by: ᴜɴᴋɴᴡᴏɴ <joe@sourcegraph.com>
Co-authored-by: Geoffrey Gilmore <geoffrey@sourcegraph.com>