ssbc: Server side resolve repos (#24222)
This PR moves resolving repositories to run over for a given Batch Spec into the backend. It exposes it as an API, so it can be previewed.
We use streaming search internally for this. Therefor, we need to expose it on the internal API, too. To make our lives a bit easier, we add RepositoryID to the streaming search payloads.