internal/repos: Better name for argument to EnqueueSingleSyncJob (#24540)
This commit renames the arg id to extSvcID in EnqueueSingleSyncJob.

I was reading this section of the code today and it wasn't immediately
obvious to me what the id referred to. This method is invoked from
repos.Syncer.TriggerExternalServiceSync and it is obvious if one looks
at the call site but not when looking independently at the method
definition.