search: remove redundant loops over repos (#24584)
* search: precompute repo set

Traces show a gap between repo resolution and the start of
`doRepoSearch`. With this change we precompute the set of repos during
an earlier loop in resolveRepositories.

* search: return all indexable repos from searchableRepos

Since we now only call resolveRepositories for diff/commit search and
repo search, we don't need to filter down searchableRepositories to
indexed only. This was necessary when we also used it for global file / path
searches, but we no longer do.

This should also improve performance of all searches on dotcom.

Co-authored-by: @stefanhengl

* fixup! lint

Co-authored-by: Tomás Senart <tsenart@gmail.com>