move store test utils into their own package (#24800)
Because testutil contained a bunch of helper functions that depend on
`internal/store`, it cannot be imported into any package that
`internal/store` depends on. This moves the testutil functions for
`internal/store` into its own subdirectory of `internal/store`, breaking
the circular dependencies and simplifying the dependency tree of
`internal/testutil`.