Let caddy wait for webpack dev server to be up (#24775)
Previously, for the first couple of seconds after starting the dev env, all requests would fail with 502 errors, because the proxied to webpack dev server isn't running yet. This config change gives it up to 60s to start up before the connection is closed with 502. NO MORE RELENTLESS SMASHING OF cmd + R.