fix diff search for git version 2.33.0 (#24803)
The latest version of git makes adding `--pickaxe-regex` and `-G` in the
same command a hard error. We are currently adding `--pickaxe-regex` to
every search that uses a regex pattern, even if we're using `-G` instead
of `-S`. This commit changes the logic to only add `--pickaxe-regex` if
we are using `-S` instead of `-G`.