Run golangci-lint in every Go project in repository (and fix all reported errors) (#24956)
I noticed that `golangci-lint` only runs over the packages belonging to `github.com/sourcegraph/sourcegraph`. It doesn't run over `lib`, `dev/sg`, etc.

This PR changes the check script we use in CI to run golangci-lint over every project in our repository with a `go.mod` (which is what we do for `go test` in CI too).

That's the first commit. The second commit fixes all of the reported errors.