cmd/gitserver: Add producer-consumer pipeline to clone repos (#24892)
In this commit we introduce a new data structure in the form of a
thread safe list to manage pending clone requests. A producer
reads from the front of the queue while new clone requests get
queued at the back. The producer writes the "job" to a channel
for a consumer to pick it up. The consumer will spawn a new
goroutine if it is able to acquire a "slot" from the cloneLimiter
- which ensures there are always going to be a maximum of
`GitMaxConcurrentClones` running concurrently at any given time
for each gitserver instance.

We've added some metrics to verify our approach in
production. While the producer-consumer pipeline would be
actively "processing" the jobs, in reality no actual cloning is
done at this time. This commit introduces the changes in a safe
way while leaving the current approach untouched. Once we have
verified with the help of the metrics that our new changes appear
to work as intended, we will enable the code path to start
cloning via the new producer-consumer pipeline and disable the
existing approach of spawning a new gorotuine for each new async
clone request.

Co-authored-by: flying-robot <71571391+flying-robot@users.noreply.github.com>
Co-authored-by: Ryan Slade <ryanslade@gmail.com>