migrations/codeintel: add API docs search indexing DB schema (#25199)
* migrations/codeintel: add API docs search indexing DB schema

This migration adds two new tables to the codeintel DB:

* `lsif_data_documentation_search_public`
* `lsif_data_documentation_search_private`

Combined, these tables will contain all of the necessary data to perform
fuzzy search over API docs data at a symbol level, globally across public
and private repositories.

For details on why we're using Postgres FTS, estimates on how well this will
scale, limitations, etc. see the architecture docs added in #25197

Note: This table reproduces some data per row, namely `repo_name` and `lang`,
which is repetitive across many rows. In a normal DB schema, this would be
advised against (as it'd be better to look this data up in another table.)
However, in this situation this (1) is a small amount of data compared to what
else is stored in this table and most importantly (2) allows for us to do some
better ranking of results as, within a single query, we can `ORDER BY` the trigram
match distance to each of these fields relatively inexpensively. As we scale the
index size up, and have more concrete queries against this table, we may revisit
this choice for better perf.

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>