temporary settings: migrate search tour from localStorage (#25276)
Add a mechanism for migrating settings stored in `localStorage` to Temporary Settings.

This is an initial proof-of-concept migrating the `'has-cancelled-onboarding-tour'` setting from `localStorage`. More settings can be added afterward but may require some refactoring depending on how they are read from and written to in `localStorage`.

This also removes any mention of the setting `'has-completed-onboarding-tour'` since it is written to but never read.