ci: add possibility to skip a step (#25498)
This is to simplify disabling a bad step as well as giving better
context. I noticed people having to do quite large changes to disable a
test, when it should be a one liner.

Documented in buildkite README. There is probably a better place in our
docs or handbook? Unsure, but then we should atleast link to it in the
README I update.