search: handle zoekt repo name being out of sync (#25628)
We have observed (one) panic in production due to repoRev being
nil. This can only happen if zoekt returns results with a repository
name different to what Sourcegraph thinks it is. This was always a
possibility (until zoekt updates indexes), but it only affected us since
we have started searching repos by ID since last week. Previously it was
by name, so the repo would be skipped and not returned.