cmd/server: use redis=~5.0 (#25669)
The redis release we depend on for this image got pulled: https://pkgs.alpinelinux.org/packages?name=redis&branch=v3.12

This is currently causing persistent build failures in main. This PR makes the constraint looser so that we can pull the dependency at least, which comes with the unfortunate tradeoff that we cannot pin the version - but it also seems like we shouldn't be pinning these versions like this anyway.