API docs: codeintel: begin indexing API docs for search (#25666)
This brings back #25206 - which was reverted in #25315 due to an issue with querying from the wrong DB (missed in dev, since codeintel and main app DBs are the same in dev this was an easy mistake to make despite me consciously trying to avoid it.)

First commit brings back the change in original state; next commits fix the actual underlying issue:

Instead of relying on the lsif_dumps table, which is in the main app DB, to locate the dumps we should delete (i.e. all rows which came from a dump for the same upload root and LSIF indexer/language), we instead introduce a dump_root column which duplicates this info on the lsif_data_documentation_search_* tables.

This is duplicative/repetitive data across rows, similar to other rows in this table. This is not optimal, but the size of the table is constrained, the duplicated data is very small, and I will improve this later once queries against these tables have solidified a bit.

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>