search: avoid the need to compute repoBranches (#25677)
We don't need to allocate repoBranches anymore. Previously we allocated
it since we sent it down as part of the zoekt query. We switched
recently to a more efficient encoding. We still used repoBranches for
mapping zoekt results back into reporevs. However, we can instead
compute that information using the same algorithm we had to compute
repoBranches.