Be more lenient about permissible batch spec structure (#25889)
Before, the editor and fromRaw endpoints would constantly throw validation errors while the user is typing in the editor, because `on:` is interpreted as `on: null`. As we don't _require_ `on`, this is just as valid, though. Now, this won't cause schema validation errors anymore. Also, this makes it easier to comment out parts of the spec without breaking validation.
I had to update the schema generator a bit, so it would still generate correct go code. Therefor, I pinned the new version in the gen script, so there won't be any flakes in CI.