commit/diff search: track matched file diffs for better merge (#25967)
This modifies our matching code to track which file diffs were matched so we can merge them correctly. This allows us to run something like file:a b and only return diffs that contain file diffs that both match file:a and match content b. Previously, this would be interpreted as "(diff contains a file diff that matches file:a) or (diff contains a file diff with content modification matching b)".