propagate database.DB up to NewSchema (#26847)
This commit propagates the new database.DB up the call chain from my test usage of it in RepoResolver all the way up the direct call chain to NewSchema. This allows me to start propagating it back down the call tree as I update tests to avoid using the global mocks.