sg: correctly download jaeger (#26984)
Currently jaeger just doesn't start in our dev environment since it
used "version" as an envvar. Additionally, we didn't notice this was
broken since "-e" bash option was not set. This is now set. We
should consider it for all shell commands in sg.

The log level is also lowered for Jaeger, since it is quiet spammy.