authz: disable p4 repo-centric sync with sub-repo perms (#27091)
Adds `authz. ErrUnimplemented`, and has `FetchRepoPerms` for a sub-repo-perms-enabled Perforce provider return it as a no-op. This allows us to safely assume that we will not accidentally grant access to a repo before sub-repo permissions are sync'd via user-centric sync.

In the long term we might want to implement sub-repo perms for repo-centric sync, but for now this is more straight-forward.