authz: Locally cache whether sub-repo perms enabled (#27124)
This change use conf.Watch to subscribe to config changes and locally
caches whether sub-repo perms are enabled at the package level.

This change is mostly needed since we don't trust the performance
implications of constantly calling conf.Get() in tight loops. In theory
it's cached, but we're unsure of how this behaves when a change occurs.