remove most uses of ensureStore (#27140)
This removes most uses of `ensureStore` methods on our stores. These
were necessary when we had global stores to make sure that we
initialized the stores _after_ `dbconn.Global` was initialized. Now that
the store types are private, they must be created through their
constructor, which requires a `dbutil.DB`.

The only store that we could not safely remove `ensureStore` from is
`repoStore` because GlobalRepos still exists and depends on this
behavior.