add docs for testing with a database dependency (#27330)
This adds some docs for my take on current best practices for testing
with a database dependency. I think the dbmock stores have matured enough that
it's worth sharing the new pattern with others, and while I'm at it, I also added
a section that explains dbtest.NewDB(), with some clarification on when to use
each package.