search: filter sub-repo perms in run.Aggregator (#27189)
Filters sub-repo perms in the results Aggregator, used by both streaming and graphql search. The entire check is omitted if the enableSubRepoPerms site config is not enabled.

Applying sub-repo permissions here is not the intended final design - tackling longer-term solutions will be tracked in #27372 upon completion of the RFC 503 project.