replace uses of ReadActionEmailQuery with ListEmailActions (#27508)
This replaces the remaining use of ReadActionEmailQuery with
ListEmailActions, and removes ReadActionEmailQuery. This also removes
one more instance of (CodeMonitorStore).Query(). Additionally,
this allows us to make scanEmails private, which is nice since
we really shouldn't be exposing database scans to external
packages.

This also gets rid of some unnecessary paging that was complicating the
changed function. We should be able to handle listing all email actions
configured for a monitor without needing to page through them.