search: only marshal config for repos that have changed (#27569)
There was a subtle but benign bug in the previous code when using the
new "only get config for what has changed" issue. We would return a 404
error for every repository that hasn't changed. This would be ignored by
zoekt and as a side effect would just cause log and metric spam.

We fix the code path by only marshalling configuration for repositories
that have changed.