WIP git-combine: parse git fast-export output
This is a mid progress prototype of parsing the output of "git
fast-export" for the purpose of rewriting paths. It successfully parses
all the output from running on the sourcegraph repository. With a little
more work it can be adjusted to just rewrite paths and pipe the input to
an output.

However, I am changing approaches. So this commit is to share context
only.