search: don't surface ctx error from repo resolution (#23296)
This adds a check to handle context errors during repo resolution.

For a very fast global search, it can happen that the context is
canceled before repo resolution is complete, in which case
resolveRepositories returns a wrapped `Context.Canceled`. So far we
returned context errors from repo resolution directly to the user.